Fix close button bug in production
